+++ service/grails-app/views/correspondence/_correspondence.groovy
@@ -1,0 +1,8 @@
+import org.olf.oa.Correspondence
+import groovy.transform.Field
+
+def should_expand = []
+
+@Field
+Correspondence correspondence
+json g.render(correspondence, [expand: should_expand])